benerin ui dan buat template observer

+++ Pomodoro/src/state/PomodoroState.java
@@ -6,10 +6,10 @@ public abstract class PomodoroState {
 
     protected Integer currentStateNumber;
 
-    protected abstract Double getTimer();
+    protected abstract int getTimer();
     protected abstract Color getColor();
     public abstract PomodoroState nextState();
-    
+     
     public PomodoroState(Integer currentStateNumber) {
         this.currentStateNumber = currentStateNumber;
     }
